+++ src/main/java/khirtech/coronadz/demo/authentication/AuthRepository.java
@@ -1,0 +1,4 @@
+package khirtech.coronadz.demo.authentication;
+
+public class AuthRepository {
+}
