+++ src/main/java/khirtech/coronadz/demo/authentication/AuthService.java
@@ -1,0 +1,4 @@
+package khirtech.coronadz.demo.authentication;
+
+public class AuthService {
+}
