+++ src/main/java/khirtech/coronadz/demo/authentication/models/AuthenticationResponse.java
@@ -1,4 +1,4 @@
-package khirtech.coronadz.demo.authetication.models;
+package khirtech.coronadz.demo.authentication.models;
 
 public class AuthenticationResponse {
