--- src/main/java/khirtech/coronadz/demo/authetication/AuthRepository.java
@@ -1,4 +1,0 @@
-package khirtech.coronadz.demo.authetication;
-
-public class AuthRepository {
-}
